Add authorization in delete & update user by login & id

+++ api/controllers/users.js
@@ -11,7 +11,7 @@ let createUser = function(req, res) {
 
 	usersServices.createUser(login, password, firstName, lastName)
 		.then((user) => {
-			let token = jwt.sign({ id: user.insertedId }, config.secret);
+			let token = jwt.sign({ id: user.insertedId, login: user.ops[0].login }, config.secret);
 
 			res.status(201);
 			res.write(JSON.stringify({ auth: true, token: token }, null, 4));
@@ -49,50 +49,82 @@ let updateUserById = function(req, res) {
 	let userId = req.swagger.params.userId.value.trim();
 	let firstName = req.swagger.params.firstName.value.trim();
 	let lastName = req.swagger.params.lastName.value.trim();
+	let token = req.headers["x-access-token"];
 
-	usersServices.updateUserById(userId, firstName, lastName)
-		.then((doc) => {
-			res.write(JSON.stringify(doc, null, 4));
-			res.end();
-		});
+	if (!token) return res.status(401).send({ auth: false, message: "No token provided." });
+
+	jwt.verify(token, config.secret, function(err, decodedUser) {
+		if (err) return res.status(500).send({ auth: false, message: "Failed to authenticate token." });
+
+		if (userId != decodedUser.id) return res.status(403).send({ auth: false, message: "Access is denied" });
+
+		usersServices.updateUserById(userId, firstName, lastName)
+			.then((doc) => {
+				res.write(JSON.stringify(doc, null, 4));
+				res.end();
+			});
+	});
 };
 
 let updateUserByLogin = function(req, res) {
 	let userLogin = req.swagger.params.userLogin.value.trim();
 	let firstName = req.swagger.params.firstName.value.trim();
 	let lastName = req.swagger.params.lastName.value.trim();
+	let token = req.headers["x-access-token"];
 
-	usersServices.updateUserByLogin(userLogin, firstName, lastName)
-		.then((doc) => {
-			res.write(JSON.stringify(doc, null, 4));
-			res.end();
-		})
-		.catch((err) => {
-			server.sendError(res, err);
-		});
+	if (!token) return res.status(401).send({ auth: false, message: "No token provided." });
+
+	jwt.verify(token, config.secret, function(err, decodedUser) {
+		if (userLogin != decodedUser.login) return res.status(403).send({ auth: false, message: "Access is denied" });
+
+		usersServices.updateUserByLogin(userLogin, firstName, lastName)
+			.then((doc) => {
+				res.write(JSON.stringify(doc, null, 4));
+				res.end();
+			})
+			.catch((err) => {
+				server.sendError(res, err);
+			});
+	});
 };
 
 let deleteUserById = function(req, res) {
 	let userId = req.swagger.params.userId.value.trim();
+	let token = req.headers["x-access-token"];
 
-	usersServices.deleteUserById(userId)
-		.then((doc) => {
-			res.write(JSON.stringify(doc, null, 4));
-			res.end();
-		});
+	if (!token) return res.status(401).send({ auth: false, message: "No token provided." });
+
+	jwt.verify(token, config.secret, function(err, decodedUser) {
+		if (err) return res.status(500).send({ auth: false, message: "Failed to authenticate token." });
+
+		if (userId != decodedUser.id) return res.status(403).send({ auth: false, message: "Access is denied" });
+
+		usersServices.deleteUserById(userId)
+			.then((doc) => {
+				res.write(JSON.stringify(doc, null, 4));
+				res.end();
+			});
+	});
 };
 
 let deleteUserByLogin = function(req, res) {
 	let userLogin = req.swagger.params.userLogin.value.trim();
+	let token = req.headers["x-access-token"];
 
-	usersServices.deleteUserByLogin(userLogin)
-		.then((doc) => {
-			res.write(JSON.stringify(doc, null, 4));
-			res.end();
-		})
-		.catch((err) => {
-			server.sendError(res, err);
-		});
+	if (!token) return res.status(401).send({ auth: false, message: "No token provided." });
+
+	jwt.verify(token, config.secret, function(err, decodedUser) {
+		if (userLogin != decodedUser.login) return res.status(403).send({ auth: false, message: "Access is denied" });
+
+		usersServices.deleteUserByLogin(userLogin)
+			.then((doc) => {
+				res.write(JSON.stringify(doc, null, 4));
+				res.end();
+			})
+			.catch((err) => {
+				server.sendError(res, err);
+			});
+	});
 };
 
 module.exports = {
